# arcanium/api/openapi/arcanium.yaml
openapi: 3.1.0
info:
  title: Arcanium API
  version: "1.0.0"
  description: >
    Arcanium's cryptographic control-plane API. Authored from the actual
    Express route handlers (arcanium/api/src/routes/*.js,
    arcanium/api/src/auth/index.js) as source of truth for current
    behavior — a capture-then-enforce exercise (Prompt 22), not a redesign.
    This is the ONE authoritative contract: the Nuxt gateway allowlist
    (arcanium/ui/server/routes/gateway/[...path].ts), the CLI request
    shapes, and docs/api.md's narrative must all be checked against this
    file, not maintained as independent parallel sources of truth.
servers:
  - url: http://localhost:3001
    description: Direct API (dev/scenario scripts)
  - url: http://localhost:3000/gateway
    description: Same-origin UI gateway (browser-facing; allowlisted subset only)
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: arc_session
      description: >
        Opaque session id issued by GET /api/v1/auth/callback. When
        ARCANIUM_AUTH_ENABLED is unset, requireSession no-ops and every
        request is treated as an estate-wide operator — see
        docs/security.md.
    serviceAccountBearer:
      type: http
      scheme: bearer
      description: >
        Reserved for a future machine-to-machine identity source
        (prompts/26_external_integration_key_lifecycle.md, not yet
        executed). Not accepted by any route in this version of the API —
        listed here only so the contract does not need a breaking change
        once it is.
  responses:
    Unauthorized:
      description: No session cookie, or the session is expired/invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: >
        Authenticated, but authorize() denied the action — wrong role,
        wrong tenant, or a scoped grant that did not match (Phase 18's
        deny-by-default matrix, auth/authorize.js).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: >
        Missing, or deliberately indistinguishable from missing for a
        cross-tenant read (Phase 18's GET /applications/:id precedent —
        a supplier-admin gets 404, not 403, for another tenant's resource,
        so existence itself isn't leaked).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Request failed field validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          description: Direct-API responses use a string; gateway responses use boolean true.
          oneOf:
            - type: string
            - type: boolean
        statusCode: { type: integer }
        statusMessage: { type: string }
        message: { type: string }
        field: { type: string, description: "Direct-API 400 responses name the offending field." }
        reason: { type: string, description: "authorize() denial reason, on 403 responses." }
        action: { type: string, description: "The authorize() action that was denied, on 403 responses." }
      required: [error]

    # ── Suppliers / Applications ─────────────────────────────────────────
...
    TransitKey:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        versions: { type: ["integer", "null"] }
        latest_version: { type: ["integer", "null"] }
        exportable: { type: boolean }
        deletion_allowed: { type: boolean }
        auto_rotate_period: { type: integer, description: "Vault seconds; 0 = disabled." }
        supports_encryption: { type: boolean }
        supports_decryption: { type: boolean }
        supports_signing: { type: boolean }
        supports_derivation: { type: boolean }
        min_decryption_version: { type: integer }
        min_encryption_version: { type: integer }
        custody: { type: string, description: "API-derived label — never inferred from the key name." }
        hsm_backed: { type: boolean }
        managed_key_name: { type: ["string", "null"] }

    # ── Provisioning jobs ─────────────────────────────────────────────────
    # Prompt 22, Deliverable 3 — status legally transitions per
    # domain/state-machines.js's JOB_STATES; the enum below lists every
    # reachable value, not just the common ones.
...
    ApprovalRecord:
      type: object
      properties:
        id: { type: string, format: uuid }
        app_id: { type: string, format: uuid }
        key_name: { type: string }
        action: { type: string }
        status: { type: string, enum: [pending, approved, rejected] }
        requester: { type: string }
        approver: { type: ["string", "null"] }
        reason: { type: ["string", "null"] }
        accessor: { type: ["string", "null"] }
        source: { type: string, enum: [manual, local, external] }
        supplier_id: { type: ["string", "null"], format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, app_id, key_name, action, status]

    # ── Evidence (Prompt 16.6/16.7) ──────────────────────────────────────
...
paths:
...
  /api/v1/keys:
    get:
      summary: Merged Transit key inventory (primary cluster + vault-hsm managed keys)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/TransitKey" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Create a Transit key
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                type: { type: string, enum: [aes256-gcm96, rsa-2048, rsa-4096, ecdsa-p256] }
              required: [name]
      responses:
        "201": { description: Created }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/keys/{name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    get:
      summary: One key's metadata (primary cluster, falling back to vault-hsm)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransitKey" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/keys/{name}/rotate:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      summary: Rotate a Transit key (architect/operator/limited supplier-admin)
      responses:
        "200": { description: Rotated synchronously }
        "202": { description: Queued }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/keys/{name}/rewrap:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      summary: Rewrap ciphertext to the latest key version — CIPHERTEXT ONLY, operator-only
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { ciphertext: { type: string, pattern: "^vault:v" } }
              required: [ciphertext]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { ciphertext: { type: string } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/keys/{name}/destroy:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      summary: Request governed destruction — records an approval, does not delete
      responses:
        "202":
          description: Approval recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  approval: { $ref: "#/components/schemas/ApprovalRecord" }
                  message: { type: string }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: No application registered to attach the destroy request to }

  # ── PKI ───────────────────────────────────────────────────────────────
...
  /api/v1/suppliers/{id}/keys:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: This supplier's Transit key inventory (namespace-scoped)
      responses:
        "200":
          description: "OK (empty array on a 403 namespace boundary — by design, not an error)"
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }

  # ── Teams (Prompt 27) ────────────────────────────────────────────────
...